refactor(cli): isolate thread resolution owner
