+++ loopx/cli_commands/registry_admin.py
@@ -32,10 +32,7 @@ from ..state_migration import (
     render_state_migration_markdown,
 )
 from ..thread_agent_binding import (
-    THREAD_BINDING_RESOLUTION_SCHEMA_VERSION,
-    ThreadBindingRequestError,
     bind_thread_agent_in_registry,
-    resolve_registry_thread_agent_binding,
     resolve_thread_agent_binding,
     unbind_thread_agent_in_registry,
 )
@@ -43,11 +40,17 @@ from ..upgrade import build_upgrade_plan
 from .goal_lifecycle import handle_goal_lifecycle_command, register_goal_lifecycle_command
 from .registry_admin_configure import register_configure_goal_command
 from .registry_admin_peer import render_register_agent_markdown
+from .registry_admin_thread_resolution import (
+    REGISTRY_THREAD_RESOLUTION_COMMANDS,
+    handle_registry_thread_resolution_command,
+    register_registry_thread_resolution_command,
+)
 from .registry_authority import (
     REGISTRY_AUTHORITY_COMMANDS,
     handle_registry_authority_command,
     register_registry_authority_commands,
 )
+from .support_control_registry import explicit_global_registry
 
 PrintPayload = Callable[
     [dict[str, object], str, Callable[[dict[str, object]], str]],
@@ -66,32 +69,7 @@ REGISTRY_ADMIN_COMMANDS = {
     "uninstall-project",
     "sync-global",
     "migrate-state",
-} | REGISTRY_AUTHORITY_COMMANDS
-
-
-def _render_thread_binding_resolution_markdown(payload: dict[str, object]) -> str:
-    lines = [
-        "# LoopX Host Thread Binding",
-        "",
-        f"- ok: `{payload.get('ok')}`",
-        f"- status: `{payload.get('status')}`",
-        f"- host_surface: `{payload.get('host_surface')}`",
-        f"- thread_id: `{payload.get('thread_id')}`",
-    ]
-    if payload.get("goal_id"):
-        lines.append(f"- goal_id: `{payload.get('goal_id')}`")
-    if payload.get("agent_id"):
-        lines.append(f"- agent_id: `{payload.get('agent_id')}`")
-    if payload.get("error_kind"):
-        lines.append(f"- error_kind: `{payload.get('error_kind')}`")
-    if payload.get("error"):
-        lines.extend(["", str(payload["error"])])
-    return "\n".join(lines)
-
-
-def explicit_global_registry(runtime_root_arg: str | None) -> Path:
-    runtime_root = Path(runtime_root_arg).expanduser() if runtime_root_arg else DEFAULT_RUNTIME_ROOT
-    return global_registry_path(runtime_root)
+} | REGISTRY_AUTHORITY_COMMANDS | REGISTRY_THREAD_RESOLUTION_COMMANDS
 
 
 def _registry_goal(path: Path, goal_id: str) -> dict[str, object]:
@@ -422,16 +400,7 @@ def register_registry_admin_commands(subparsers: argparse._SubParsersAction) -> 
         help="Write the source registry and sync it globally. Without this flag, preview only.",
     )
 
-    resolve_thread_parser = subparsers.add_parser(
-        "resolve-agent-thread",
-        help="Resolve one exact host thread across the current project registry.",
-    )
-    resolve_thread_parser.add_argument(
-        "--thread-id", required=True, help="Stable opaque host thread id."
-    )
-    resolve_thread_parser.add_argument(
-        "--host-surface", required=True, help="Exact host surface token."
-    )
+    register_registry_thread_resolution_command(subparsers)
 
     bind_thread_parser = subparsers.add_parser(
         "bind-agent-thread",
@@ -750,37 +719,13 @@ def handle_registry_admin_command(
         print_payload(payload, args.format, render_register_agent_markdown)
         return 0 if payload.get("ok") else 1
 
-    if args.command == "resolve-agent-thread":
-        try:
-            payload = resolve_registry_thread_agent_binding(
-                registry_path=registry_path,
-                host_surface=args.host_surface,
-                thread_id=args.thread_id,
-            )
-        except Exception as exc:
-            invalid_request = isinstance(exc, ThreadBindingRequestError)
-            payload = {
-                "ok": False,
-                "schema_version": THREAD_BINDING_RESOLUTION_SCHEMA_VERSION,
-                "host_surface": None if invalid_request else args.host_surface,
-                "thread_id": None if invalid_request else args.thread_id,
-                "status": "unavailable",
-                "goal_id": None,
-                "agent_id": None,
-                "matches": [],
-                "error_kind": (
-                    "thread_agent_binding_invalid_request"
-                    if invalid_request
-                    else "thread_agent_binding_resolution_failed"
-                ),
-                "error": (
-                    "thread binding request is invalid"
-                    if invalid_request
-                    else "thread binding authority could not be read"
-                ),
-            }
-        print_payload(payload, args.format, _render_thread_binding_resolution_markdown)
-        return 0 if payload.get("ok") else 1
+    thread_resolution_result = handle_registry_thread_resolution_command(
+        args,
+        registry_path=registry_path,
+        print_payload=print_payload,
+    )
+    if thread_resolution_result is not None:
+        return thread_resolution_result
 
     if args.command in {"bind-agent-thread", "unbind-agent-thread"}:
         try:
